+++ packages/solana-wallet-snap/src/core/services/assets/index.ts
@@ -1,4 +1,2 @@
-export * from './adapters/SnapAssetsAdapter';
-export * from './AssetsRepository';
 export * from './AssetsService';
 export * from './TokenHelper';